lolcensus/requirements.txt
arrow==0.17.0
asgiref==3.3.1
cassiopeia==4.0.14
certifi==2020.12.5
chardet==4.0.0
coverage==5.3.1
datapipelines==1.0.7
decorator==4.4.2
Django==3.1.5
django-cassiopeia==2.1.0
django-environ==0.4.5
django-mathfilters==1.0.0
django-redis==4.12.1
environ==1.0
idna==2.10
merakicommons==1.0.9
networkx==2.5
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
django-recaptcha==2.0.6